Backend/staiged-backendd/pyproject.toml
[tool.poetry]
name = "staiged-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pyaudio = "*"
thefuzz = "*"
ocrmypdf = "*"
pypdfium2 = "*"
numpy = "*"
pypdf = "*"
reportlab = "*"
fitz = "*"
pymupdf = "*"
faster-whisper = "*"
loguru = "*"
librosa = "*"
paho-mqtt = "^1.6.1"
torch = "^2.2.0"
torchvision = "^0.17.0"
opencv-python = "^4.9.0.80"
pillow = "^10.2.0"
ultralytics = "^8.1.19"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
